requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
antlr4-python3-runtime==4.9.3
anyio==4.12.0
asteroid-filterbanks==0.4.0
attrs==25.4.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
docopt==0.6.2
einops==0.8.1
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.24.7
HyperPyYAML==1.2.2
idna==3.11
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
kiwisolver==1.4.9
lightning==2.6.0
lightning-utilities==0.15.2
llvmlite==0.45.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.0
networkx==3.6
numba==0.62.1
numpy==2.3.5
omegaconf==2.3.0
openai-whisper==20240930
optuna==4.6.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
primePy==1.3
propcache==0.4.1
protobuf==6.33.1
pyannote.audio==3.4.0
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
pytorch-metric-learning==2.9.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
scikit-learn==1.7.2
scipy==1.16.3
semver==3.0.4
sentencepiece==0.2.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
soundfile==0.13.1
speechbrain==1.0.3
SQLAlchemy==2.0.44
sympy==1.14.0
tabulate==0.9.0
tensorboardX==2.6.4
threadpoolctl==3.6.0
tiktoken==0.12.0
torch==2.4.1
torch-audiomentations==0.12.0
torch_pitch_shift==1.2.5
torchaudio==2.4.1
torchmetrics==1.8.2
tqdm==4.67.1
typer==0.20.0
typer-slim==0.20.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
yarl==1.22.0
zipp==3.23.0
